anndata-dev compatibility

+++ scanpy/datasets/_datasets.py
@@ -180,7 +180,7 @@ def paul15() -> ad.AnnData:
         clusters = f['cluster.id'][()].flatten().astype(int)
         infogenes_names = f['info.genes_strings'][()].astype(str)
     # each row has to correspond to a observation, therefore transpose
-    adata = ad.AnnData(X.transpose(), dtype=X.dtype)
+    adata = ad.AnnData(X.transpose(), dtype=np.float32)
     adata.var_names = gene_names
     adata.row_names = cell_names
     # names reflecting the cell type identifications from the paper
